+++ in_definite/src/rules/mod.rs
@@ -1,3 +1,3 @@
-pub mod rule_acronymns;
+pub mod rule_acronyms;
 pub mod rule_exceptions;
 pub mod rule_numbers;