+++ in_definite/src/rules/rule_numbers.rs
@@ -3,27 +3,24 @@ use crate::string_helper;
 
 // ref: https://github.com/tandrewnichols/indefinite/blob/master/lib/rules/numbers.js
 pub fn is_number(word: &str) -> bool {
-    string_helper::is_match(word, r"^([0-9,]+)")
+    string_helper::get_first_letter(word).is_numeric()
 }
 
 pub fn is_an_for_number(word: &str, options: &Options) -> bool {
     let mut is_an = false;
 
-    if string_helper::is_match(word, r"^(11|8|18)") {
-        let starts_with_11_or_18 = string_helper::is_match(word, r"^(11|18)");
-
-        // If the number starts with 11 or 18 and is of length 4,
-        // the pronunciation is ambiguous so check opts.numbers to see
-        // how to render it. Otherwise, if it starts with 11 or 18
-        // and has 2, 5, 8, 11, etc. digits, use 'an.' Finally, if it
-        // starts with an 8, use 'an.' For everything else, use 'a.'
-        if starts_with_11_or_18 && word.len() == 4 {
-            is_an = options.are_numbers_colloquial;
-        } else if starts_with_11_or_18 && (word.len() - 2) % 3 == 0 {
-            is_an = true;
-        } else {
-            is_an = word.starts_with('8');
-        }
+    // If the number starts with 11 or 18 and is of length 4,
+    // the pronunciation is ambiguous so check opts.numbers to see
+    // how to render it. Otherwise, if it starts with 11 or 18
+    // and has 2, 5, 8, 11, etc. digits, use 'an.' Finally, if it
+    // starts with an 8, use 'an.' For everything else, use 'a.'
+    let starts_with_11_or_18 = word.starts_with("11") || word.starts_with("18");
+    if starts_with_11_or_18 && word.len() == 4 {
+        is_an = options.are_numbers_colloquial;
+    } else if starts_with_11_or_18 && (word.len() - 2) % 3 == 0 {
+        is_an = true;
+    } else if word.starts_with('8') {
+        is_an = true;
     }
 
     is_an